extract base class for step handlers
